Rename app from 'tasks' to 'manager'
