+++ task_manager/settings.py
@@ -38,7 +38,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "django_extensions",
-    "tasks"
+    "manager"
 ]
 
 MIDDLEWARE = [
@@ -127,4 +127,4 @@ STATICFILES_DIRS = [
 
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
 
-AUTH_USER_MODEL = "tasks.Worker"
+AUTH_USER_MODEL = "manager.Worker"
