+++ task_manager/urls.py
@@ -20,5 +20,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("tasks/", include("tasks.urls", namespace="tasks")),
+    path("manager/", include("manager.urls", namespace="manager")),
 ]
