+++ templates/includes/sidebar.html
@@ -1,5 +1,5 @@
-<a href="{% url 'tasks:index' %}" class="btn btn-secondary">Home</a>
-<a href="{% url 'tasks:worker-list' %}" class="btn btn-secondary">Workers</a>
+<a href="{% url 'manager:index' %}" class="btn btn-secondary">Home</a>
+<a href="{% url 'manager:worker-list' %}" class="btn btn-secondary">Workers</a>
 <a href="#" class="btn btn-secondary">Projects</a>
 <a href="#" class="btn btn-secondary">Tasks</a>
